fixed issue with hexahedron intersection test
found a problem with 5-tetrahedron decomposition: since it triangulates the upper
faces differently from the lower face, the face between two neighboring cells
may seem to have two different shapes depending on whether it's being considered as
the lower face of the upper cell or the upper face of the lower cell. this can lead
to an apparent gap between the two cells if the face is not planar. switched to
6-tetrahedron decomposition which does not have this behavior