improvements to overlap hole cutting and more-than-two grid interfaces
still working on a way to do this correctly automatically, so for now users need
to specify a number indicating the amount of overlap to leave between each pair
of grids. also, in order to correctly detect orphan points, complete information
about the boundaries is needed (including non-overlapping boundaries, which
weren't needed before). coming shortly will be an option to detect non-overlapping
boundaries automatically so the user doesn't need to manually specify them